+++ src/main/java/com/restful/api/response/ResponseService.java
@@ -7,28 +7,6 @@ import java.util.List;
 @Component
 public class ResponseService {
 
-    // enum으로 API 요청 결과에 대한 code, message를 정의한다.
-    public enum CommonResponse {
-        SUCCESS(0, "성공했습니다."),
-        FAIL(-1, "실패했습니다.");
-
-        int code;
-        String message;
-
-        CommonResponse(int code, String message) {
-            this.code = code;
-            this.message = message;
-        }
-
-        public int getCode() {
-            return code;
-        }
-
-        public String getMessage() {
-            return message;
-        }
-    }
-
     // 단일 결과를 처리하는 메서드
     public <T> SingleResult<T> getSingleResult(T data) {
         SingleResult<T> result = new SingleResult<>();
